Break our TreeRow into its own separate file.
#254 When adding a goal using drag and drop, the title isn't editable.
